# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is URL shortener API",
        "title": "Shortener API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/RouSage/shortener/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "paths": {
...
        "/v1/urls": {
            "get": {
                "description": "Retrieves a paginated list of URLs created by the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URLs"
                ],
                "summary": "Get User URLs",
                "parameters": [
                    {
                        "maximum": 10000,
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of user URLs",
                        "schema": {
                            "$ref": "#/definitions/server.PaginatedUserURLs"
                        }
                    },
                    "400": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "description": "Creates a shortened URL. Authenticated users can provide a custom short code (5-16 characters). Otherwise, a random code is generated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URLs"
                ],
                "summary": "Create Short URL",
                "parameters": [
                    {
                        "description": "URL and optional custom short code",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.CreateShortUrlDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created short URL",
                        "schema": {
                            "$ref": "#/definitions/repository.Url"
                        }
                    },
                    "400": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPValidationError"
                        }
                    },
                    "403": {
                        "description": "Custom short codes require authentication",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Short code already taken or validation failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/v1/urls/{code}": {
            "get": {
                "description": "Retrieves the original long URL for a given short code. Checks cache first, then database.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URLs"
                ],
                "summary": "Get Long URL",
                "parameters": [
                    {
                        "maxLength": 16,
                        "type": "string",
                        "description": "Short code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "longUrl",
                        "schema": {
                            "$ref": "#/definitions/server.GetLongUrlResponse"
                        }
                    },
                    "400": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPValidationError"
                        }
                    },
                    "404": {
                        "description": "Short URL not found",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "description": "Deletes a short URL owned by the authenticated user. Also removes it from cache.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URLs"
                ],
                "summary": "Delete Short URL",
                "parameters": [
                    {
                        "maxLength": 16,
                        "type": "string",
                        "description": "Short code to delete",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content - URL successfully deleted"
                    },
                    "400": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Short URL not found or not owned by user",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "definitions": {
        "appvalidator.ValidationError": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "repository.Url": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isCustom": {
                    "type": "boolean"
                },
                "longUrl": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "server.CreateShortUrlDTO": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "shortCode": {
                    "type": "string",
                    "maxLength": 16,
                    "minLength": 5
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "server.GetLongUrlResponse": {
            "type": "object",
            "properties": {
                "longUrl": {
                    "type": "string"
                }
            }
        },
        "server.HTTPError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
        "server.HTTPValidationError": {
            "type": "object",
            "properties": {
                "errors": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/appvalidator.ValidationError"
                        }
                    ],
                    "example": {
                        "{\"field\"": "\"field error message\"}"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "server.PaginatedUserURLs": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.URLResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/server.Pagination"
                }
            }
        },
        "server.Pagination": {
            "type": "object",
            "properties": {
                "hasNext": {
                    "type": "boolean"
                },
                "hasPrevious": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalItems": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "server.URLResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isCustom": {
                    "type": "boolean"
                },
                "longUrl": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}